Only secure APIs and not public
